requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
easyocr>=1.7.0
python-docx>=0.8.11
python-pptx>=0.6.21
Pillow>=10.0.0
magic-pdf[full]>=0.6.0
google-generativeai==0.5.0

# Additional dependencies for proper functionality
typing-extensions>=4.0.0